Working Obstacles & Destroy Clones
Obstacles: Remove 5 points from player. If the score goes below 0, the text appears as 0 to prevent the score going into the negatives.
Clone objects are destroyed once they go below the screen and are out of the view of the player. To prevent lag and too many objects on the scene at once.